+++ src/components/common/modal/ModalWrapper.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, useEffect, useRef } from 'react';
 import { createPortal } from 'react-dom';
 import classNames from 'classnames/bind';
-import styles from './Modal.module.scss';
+import styles from './ModalWrapper.module.scss';
 
 const cx = classNames.bind(styles);
