+++ src/components/common/modal/confirm-modal/ConfirmModal.tsx
@@ -1,7 +1,7 @@
 import classNames from 'classnames/bind';
 import ButtonPair from '@/components/common/button-pair/ButtonPair';
 import Button from '@/components/common/button/Button';
-import ModalWrapper from '@/components/common/modal/Modal';
+import ModalWrapper from '@/components/common/modal/ModalWrapper';
 import styles from './ConfirmModal.module.scss';
 
 const cx = classNames.bind(styles);
